Update install.sh
added confirmation to credentials in case of mistype

+++ install.sh
@@ -68,19 +68,31 @@ if [[ ${ACCEPTED,,} =~ ^[y] ]]; then
      wget -q https://raw.githubusercontent.com/Illusionist2732/HP-ILO-Fan-Control-ML350/main/Files/$AUTOFANFILE -O autofan.sh
     #wget -q https://raw.githubusercontent.com/That-Guy-Jack/HP-ILO-Fan-Control/main/Files/$AUTOFANFILE -O autofan.sh
 
-    read -rep 'Enter iLO Username: ' ILOUSERNAME
-    read -rep 'Enter iLO Password: ' ILOPASSWORD
-    read -rep 'Enter iLO IP/hostname: ' ILOHOST
-    sed -ri "s/your username/$ILOUSERNAME/" autofan.sh
-    sed -ri "s/your password/$ILOPASSWORD/" autofan.sh
-    sed -ri "s/your ilo ip/$ILOHOST/" autofan.sh
-    mv autofan.sh /
-
-    echo -e "\e[92mDone! Please visit the GitHub page to follow the instructions!\e[0m"
-    echo -e "\e[1\https://github.com/That-Guy-Jack/HP-ILO-Fan-Control\e[0m"
-
-else
-    echo -e "\e[31m:( exiting\e[0m"
-    exit 0
-
+while true; do
+        read -rep 'Enter iLO Username: ' ILOUSERNAME
+        read -rep 'Enter iLO Password: ' ILOPASSWORD
+        read -rep 'Enter iLO IP/hostname: ' ILOHOST
+           
+        echo
+        echo "Please confirm the following details:"
+        echo "  Username: $ILOUSERNAME"
+        echo "  Password: $ILOPASSWORD"
+        echo "  Host:     $ILOHOST"
+        echo
+            read -rep "Are These Correct? (y/N): " ACCEPTED
+                if [[ ${ACCEPTED,,} =~ ^[y] ]]; then
+                    sed -ri "s/your username/$ILOUSERNAME/" autofan.sh
+                    sed -ri "s/your password/$ILOPASSWORD/" autofan.sh
+                    sed -ri "s/your ilo ip/$ILOHOST/" autofan.sh
+                        mv autofan.sh /
+                    echo -e "\e[92mDone! Please visit the GitHub page to follow the instructions!\e[0m"
+                    echo -e "\e[1\https://github.com/That-Guy-Jack/HP-ILO-Fan-Control\e[0m"
+                    echo -e "\e[1\You may need to edit /bin/bash /autofan.sh To change the credentials if they didn't save properly.\e[0m"
+                    break
+                else
+            echo -e "Try entering the Credentials again."
+        fi
+done
+echo -e "\e[31m:( exiting\e[0m"
+exit 0
 fi
